[xs_rviz] Rename rviz_interface to xs_rviz
Changed the name of the interbotix_rviz_interface_package to
interbotix_xs_rviz because the application within this package are
specific to the X-Series robots. In the future, we may add packages
that work with all robot in IRROS, so that package name will be
reserved.

Added a README to xs_rviz describing the package structure and usage.
