+++ interbotix_xs_toolbox/interbotix_xs_rviz/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.0.2)
-project(interbotix_rviz_interface)
+project(interbotix_xs_rviz)
 catkin_package()
 find_package(catkin REQUIRED COMPONENTS
     rviz
@@ -33,7 +33,7 @@ add_definitions(-DQT_NO_KEYWORDS)
 ## Here we specify the list of source files.
 ## The generated MOC files are included automatically as headers.
 set(SRC_FILES
-  src/torque_enable_panel.cpp include/interbotix_rviz_interface/torque_enable_panel.hpp
+  src/torque_enable_panel.cpp include/interbotix_xs_rviz/torque_enable_panel.hpp
 )
 
 add_library(
